+++ tests/test_datatype.py
@@ -25,7 +25,7 @@ def test_package_encode_decode():
             value = value.replace(old, new)
         return value
 
-    process_package_encoded = dict(process)["package"]
+    process_package_encoded = process.params()["package"]
     assert "cwl.Version" not in process_package_encoded
     assert _replace_specials("cwl.Version") in process_package_encoded
     assert "$namespaces" not in process_package_encoded
@@ -38,4 +38,4 @@ def test_package_encode_decode():
     assert _replace_specials("$format") in process_package_encoded["outputs"]["output"]
     assert "$namespace" not in process_package_encoded["executionUnits"][0]["unit"]
     assert _replace_specials("$namespace") in process_package_encoded["executionUnits"][0]["unit"]
-    assert package == process.package  # gets decoded
+    assert package == process.package, "package obtained from the process method should be the original decoded version"